Constrain GPS coords to lat:[-90,90], lng:[-180,180]
Instead of displaying GPS coordinates out of the above ranges when
leaflet wraps the map, just wrap the coordinates with it.

Delayed the rounding of GPS locations until display, so the conversions
still get the full GPS coords.

Changed some bounding stuff for UTM, to not error out in some
cases.
This does do stuff like subtracting a small number from -180 and similar
things, because the included UTM math can't deal with that properly.